third_party/blink/web_tests/fast/media/media-query-list-listener-expected.txt

Media query listeners should fire on changes to matches status.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mediaList.matches is true
PASS callbackCount is 0
PASS mediaList.matches is false
PASS mediaList.matches is mediaListArgument.matches
PASS mediaList.media is mediaListArgument.media
PASS mediaList.matches is false
PASS callbackCount is 1
PASS mediaList.matches is false
PASS callbackCount is 1
PASS mediaList.matches is true
PASS mediaList.matches is mediaListArgument.matches
PASS mediaList.media is mediaListArgument.media
PASS mediaList.matches is true
PASS callbackCount is 2
PASS mediaList.matches is false
PASS callbackCount is 2
PASS successfullyParsed is true

TEST COMPLETE
